refactor and add new postsubmit job
