+++ ci/prow/make_config.go
@@ -25,7 +25,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
-	"math"
 	"os"
 	"path"
 	"regexp"
@@ -49,19 +48,6 @@ const (
 	// ##########################################################
 	// generalProwConfig contains config-wide definitions.
 	generalProwConfig = "prow_config_header.yaml"
-
-	// presubmitJob is the template for presubmit jobs.
-	presubmitJob = "prow_presubmit_job.yaml"
-
-	// presubmitGoCoverageJob is the template for go coverage presubmit jobs.
-	presubmitGoCoverageJob = "prow_presubmit_gocoverate_job.yaml"
-
-	// goCoveragePostsubmitJob is the template for the go postsubmit coverage job.
-	goCoveragePostsubmitJob = "prow_postsubmit_gocoverage_job.yaml"
-
-	// perfOpsPostsubmitJob is the template for the perf cluster operations
-	// postsubmit job
-	perfOpsPostsubmitJob = "prow_postsubmit_perfops_job.yaml"
 
 	// pluginsConfig is the template for the plugins YAML file.
 	pluginsConfig = "prow_plugins.yaml"
@@ -129,21 +115,6 @@ type baseProwJobTemplateData struct {
 // ####################################################################################################
 // ################ data definitions that are used for the prow config file generation ################
 // ####################################################################################################
-// presubmitJobTemplateData contains data about a presubmit Prow job.
-type presubmitJobTemplateData struct {
-	Base                 baseProwJobTemplateData
-	PresubmitJobName     string
-	PresubmitPullJobName string
-	PresubmitPostJobName string
-	PresubmitCommand     []string
-}
-
-// postsubmitJobTemplateData contains data about a postsubmit Prow job.
-type postsubmitJobTemplateData struct {
-	Base              baseProwJobTemplateData
-	PostsubmitJobName string
-	PostsubmitCommand []string
-}
 
 // sectionGenerator is a function that generates Prow job configs given a slice of a yaml file with configs.
 type sectionGenerator func(string, string, yaml.MapSlice)
@@ -206,57 +177,6 @@ var (
 
 	projNameRegex = regexp.MustCompile(`.+-[0-9\.]+$`)
 )
-
-// Generate cron string based on job type, offset generated from jobname
-// instead of assign random value to ensure consistency among runs,
-// timeout is used for determining how many hours apart
-func generateCron(jobType, jobName string, timeout int) string {
-	getUTCtime := func(i int) int { return i + 7 }
-	// Sums the ascii valus of all letters in a jobname,
-	// this value is used for deriving offset after hour
-	var sum float64
-	for _, c := range jobType + jobName {
-		sum += float64(c)
-	}
-	// Divide 60 minutes into 6 buckets
-	bucket := int(math.Mod(sum, 6))
-	// Offset in bucket, range from 0-9, first mod with 11(a random prime number)
-	// to ensure every digit has a chance (i.e., if bucket is 0, sum has to be multiply of 6,
-	// so mod by 10 can only return even number)
-	offsetInBucket := int(math.Mod(math.Mod(sum, 11), 10))
-	minutesOffset := bucket*10 + offsetInBucket
-	// Determines hourly job inteval based on timeout
-	hours := int((timeout+5)/60) + 1 // Allow at least 5 minutes between runs
-	hourCron := fmt.Sprintf("%d * * * *", minutesOffset)
-	if hours > 1 {
-		hourCron = fmt.Sprintf("%d */%d * * *", minutesOffset, hours)
-	}
-	dayCron := fmt.Sprintf("%d %%d * * *", minutesOffset)    // hour
-	weekCron := fmt.Sprintf("%d %%d * * %%d", minutesOffset) // hour, weekday
-
-	var res string
-	switch jobType {
-	case "continuous", "custom-job", "auto-release": // Every hour
-		res = fmt.Sprintf(hourCron)
-	case "branch-ci": // Every day 1-2 PST
-		res = fmt.Sprintf(dayCron, getUTCtime(1))
-	case "nightly": // Every day 2-3 PST
-		res = fmt.Sprintf(dayCron, getUTCtime(2))
-	case "dot-release": // Every Tuesday 2-3 PST
-		res = fmt.Sprintf(weekCron, getUTCtime(2), 2)
-	case "latency": // Every day 1-2 PST
-		res = fmt.Sprintf(dayCron, getUTCtime(1))
-	case "performance": // Every day 1-2 PST
-		res = fmt.Sprintf(dayCron, getUTCtime(1))
-	case "performance-mesh": // Every day 3-4 PST
-		res = fmt.Sprintf(dayCron, getUTCtime(3))
-	case "webhook-apicoverage": // Every day 2-3 PST
-		res = fmt.Sprintf(dayCron, getUTCtime(2))
-	default:
-		log.Printf("job type not supported for cron generation '%s'", jobName)
-	}
-	return res
-}
 
 // Yaml parsing helpers.
 
@@ -683,123 +603,6 @@ func parseBasicJobConfigOverrides(data *baseProwJobTemplateData, config yaml.Map
 	}
 }
 
-// generatePresubmit generates all presubmit job configs for the given repo and configuration.
-func generatePresubmit(title string, repoName string, presubmitConfig yaml.MapSlice) {
-	var data presubmitJobTemplateData
-	data.Base = newbaseProwJobTemplateData(repoName)
-	data.Base.Command = presubmitScript
-	data.Base.GoCoverageThreshold = 50
-	jobTemplate := readTemplate(presubmitJob)
-	repoData := repositoryData{Name: repoName, EnableGoCoverage: false, GoCoverageThreshold: data.Base.GoCoverageThreshold}
-	isMonitoredJob := false
-	generateJob := true
-	for i, item := range presubmitConfig {
-		switch item.Key {
-		case "build-tests", "unit-tests", "integration-tests":
-			if !getBool(item.Value) {
-				return
-			}
-			jobName := getString(item.Key)
-			data.PresubmitJobName = data.Base.RepoNameForJob + "-" + jobName
-			// Use default arguments if none given.
-			if len(data.Base.Args) == 0 {
-				data.Base.Args = []string{"--" + jobName}
-			}
-			if item.Key == "integration-tests" || item.Key == "unit-tests" {
-				isMonitoredJob = true
-			}
-			addVolumeToJob(&data.Base, "/etc/repoview-token", "repoview-token", true, "")
-		case "go-coverage":
-			if !getBool(item.Value) {
-				return
-			}
-			jobTemplate = readTemplate(presubmitGoCoverageJob)
-			data.PresubmitJobName = data.Base.RepoNameForJob + "-go-coverage"
-			data.Base.Image = coverageDockerImage
-			data.Base.ServiceAccount = ""
-			repoData.EnableGoCoverage = true
-			addVolumeToJob(&data.Base, "/etc/covbot-token", "covbot-token", true, "")
-		case "custom-test":
-			data.PresubmitJobName = data.Base.RepoNameForJob + "-" + getString(item.Value)
-		case "go-coverage-threshold":
-			data.Base.GoCoverageThreshold = getInt(item.Value)
-			repoData.GoCoverageThreshold = data.Base.GoCoverageThreshold
-		case "repo-settings":
-			generateJob = false
-		default:
-			continue
-		}
-		// Knock-out the item, signalling it was already parsed.
-		presubmitConfig[i] = yaml.MapItem{}
-	}
-	repositories = append(repositories, repoData)
-	parseBasicJobConfigOverrides(&data.Base, presubmitConfig)
-	if !generateJob {
-		return
-	}
-	data.PresubmitCommand = createCommand(data.Base)
-	data.PresubmitPullJobName = "pull-" + data.PresubmitJobName
-	data.PresubmitPostJobName = "post-" + data.PresubmitJobName
-	if data.Base.ServiceAccount != "" {
-		addEnvToJob(&data.Base, "GOOGLE_APPLICATION_CREDENTIALS", data.Base.ServiceAccount)
-		addEnvToJob(&data.Base, "E2E_CLUSTER_REGION", "us-central1")
-	}
-	if isMonitoredJob {
-		addMonitoringPubsubLabelsToJob(&data.Base, data.PresubmitPullJobName)
-	}
-	addExtraEnvVarsToJob(extraEnvVars, &data.Base)
-	configureServiceAccountForJob(&data.Base)
-	jobName := data.PresubmitPullJobName
-	executeJobTemplateWrapper(repoName, &data, func(data interface{}) {
-		executeJobTemplate("presubmit", jobTemplate, title, repoName, jobName, true, data)
-	})
-	// TODO(adrcunha): remove once the coverage-dev job isn't necessary anymore.
-	// Generate config for pull-knative-serving-go-coverage-dev right after pull-knative-serving-go-coverage
-	if data.PresubmitPullJobName == "pull-knative-serving-go-coverage" {
-		data.PresubmitPullJobName += "-dev"
-		data.Base.AlwaysRun = false
-		data.Base.Image = strings.Replace(data.Base.Image, "coverage:latest", "coverage-dev:latest", -1)
-		data.Base.Image = strings.Replace(data.Base.Image, "coverage-go112:latest", "coverage-dev:latest", -1)
-		template := strings.Replace(readTemplate(presubmitGoCoverageJob), "(all|", "(", 1)
-		executeJobTemplate("presubmit", template, title, repoName, data.PresubmitPullJobName, true, data)
-	}
-}
-
-// generateGoCoveragePostsubmit generates the go coverage postsubmit job config for the given repo.
-func generateGoCoveragePostsubmit(title, repoName string, _ yaml.MapSlice) {
-	var data postsubmitJobTemplateData
-	data.Base = newbaseProwJobTemplateData(repoName)
-	data.Base.Image = coverageDockerImage
-	data.PostsubmitJobName = fmt.Sprintf("post-%s-go-coverage", data.Base.RepoNameForJob)
-	for _, repo := range repositories {
-		if repo.Name == repoName && repo.DotDev {
-			data.Base.PathAlias = "path_alias: knative.dev/" + path.Base(repoName)
-		}
-		if repo.Name == repoName && repo.Go113 {
-			data.Base.Image = getGo113ImageName(data.Base.Image)
-		}
-	}
-	data.Base.ExtraRefs = append(data.Base.ExtraRefs, "  base_ref: "+data.Base.RepoBranch)
-	if data.Base.PathAlias != "" {
-		data.Base.ExtraRefs = append(data.Base.ExtraRefs, "  "+data.Base.PathAlias)
-	}
-	
-	addExtraEnvVarsToJob(extraEnvVars, &data.Base)
-	configureServiceAccountForJob(&data.Base)
-	jobName := data.PostsubmitJobName
-	executeJobTemplateWrapper(repoName, &data, func(data interface{}) {
-		executeJobTemplate("postsubmit go coverage", readTemplate(goCoveragePostsubmitJob), title, repoName, jobName, false, data)
-	})
-	// TODO(adrcunha): remove once the coverage-dev job isn't necessary anymore.
-	// Generate config for post-knative-serving-go-coverage-dev right after post-knative-serving-go-coverage
-	if data.PostsubmitJobName == "post-knative-serving-go-coverage" {
-		data.PostsubmitJobName += "-dev"
-		data.Base.Image = strings.Replace(data.Base.Image, "coverage-go112:latest", "coverage-dev:latest", -1)
-		data.Base.Image = strings.Replace(data.Base.Image, "coverage:latest", "coverage-dev:latest", -1)
-		executeJobTemplate("presubmit", readTemplate(goCoveragePostsubmitJob), title, repoName, data.PostsubmitJobName, false, data)
-	}
-}
-
 // getProwConfigData gets some basic, general data for the Prow config.
 func getProwConfigData(config yaml.MapSlice) prowConfigTemplateData {
 	var data prowConfigTemplateData
@@ -841,16 +644,6 @@ func parseSection(config yaml.MapSlice, title string, generate sectionGenerator,
 				finalize(title, repoName, nil)
 			}
 		}
-	}
-}
-
-// generateOtherJobConfigs generates job config with the generator if new job is required for it.
-func generateOtherJobConfigs(title string, newJobNeeded newJobNeeded, generate sectionGenerator) {
-	for i := range repositories { // Keep order for predictable output.
-		if !newJobNeeded(repositories[i]) {
-			continue
-		}
-		generate(title, repositories[i].Name, nil)
 	}
 }
 
@@ -1010,7 +803,7 @@ func recursiveSBL(repoName string, data interface{}, generateOneJob func(data in
 	sb.restore(base)
 }
 
-// executeJobTemplateWrapper takes in consideration of repo settings, decides how many variants of the
+// executeJobTemplateWrapper takes in consideration of repo settings, decides how many varianats of the
 // same job needs to be generated and generates them.
 func executeJobTemplateWrapper(repoName string, data interface{}, generateOneJob func(data interface{})) {
 	var sbs []specialBranchLogic
@@ -1201,11 +994,11 @@ func collectMetaData(periodicJob yaml.MapSlice) {
 				jobDetailMap[repoName] = newJobTypes
 			}
 		}
-		addTestCoverageJobIfNeeded(&jobDetailMap, repoName)
+		addTestCoverageJobDataIfNeeded(&jobDetailMap, repoName)
 	}
 
 	// add test coverage jobs for the repos that haven't been handled
-	addRemainingTestCoverageJobs()
+	addRemainingTestCoverageJobsData()
 }
 
 // addProjAndRepoIfNeed adds the project and repo if they are new in the metaData map, then return the jobDetailMap
@@ -1229,8 +1022,8 @@ func addProjAndRepoIfNeed(projName string, repoName string) map[string][]string 
 	return jobDetailMap
 }
 
-// addTestCoverageJobIfNeeded adds test-coverage job for the repo if it has go coverage check
-func addTestCoverageJobIfNeeded(jobDetailMap *map[string][]string, repoName string) {
+// addTestCoverageJobDataIfNeeded adds test-coverage job data for the repo if it has go coverage check
+func addTestCoverageJobDataIfNeeded(jobDetailMap *map[string][]string, repoName string) {
 	if goCoverageMap[repoName] {
 		newJobTypes := append((*jobDetailMap)[repoName], "test-coverage")
 		(*jobDetailMap)[repoName] = newJobTypes
@@ -1240,8 +1033,8 @@ func addTestCoverageJobIfNeeded(jobDetailMap *map[string][]string, repoName stri
 	}
 }
 
-// addRemainingTestCoverageJobs adds test-coverage jobs for the repos that haven't been processed.
-func addRemainingTestCoverageJobs() {
+// addRemainingTestCoverageJobsData adds test-coverage jobs data for the repos that haven't been processed.
+func addRemainingTestCoverageJobsData() {
 	// handle repos that only have go coverage
 	for repoName, hasGoCoverage := range goCoverageMap {
 		if hasGoCoverage {
@@ -1370,9 +1163,11 @@ func main() {
 		}
 		parseSection(config, "presubmits", generatePresubmit, nil)
 		parseSection(config, "periodics", generatePeriodic, generateGoCoveragePeriodic)
-		generateOtherJobConfigs("periodics", func(repo repositoryData) bool {
-			return !repo.Processed && repo.EnableGoCoverage
-		}, generateGoCoveragePeriodic)
+		for _, repo := range repositories { // Keep order for predictable output.
+			if !repo.Processed && repo.EnableGoCoverage {
+				generateGoCoveragePeriodic("periodics", repo.Name, nil)
+			}
+		}
 		if *generateMaintenanceJobs {
 			generateCleanupPeriodicJob()
 			generateFlakytoolPeriodicJob()
@@ -1381,10 +1176,12 @@ func main() {
 			generateIssueTrackerPeriodicJobs()
 			generatePerfClusterUpdatePeriodicJobs()
 		}
-		generateOtherJobConfigs("postsubmits", func(repo repositoryData) bool {
-			return repo.EnableGoCoverage
-		}, generateGoCoveragePostsubmit)
-		generatePerfClusterReconcilePostsubmitJobs()
+		for _, repo := range repositories {
+			if repo.EnableGoCoverage {
+				generateGoCoveragePostsubmit("postsubmits", repo.Name, nil)
+			}
+			generatePerfClusterReconcilePostsubmitJob(repo.Name)
+		}
 	}
 
 	// config object is modified when we generate prow config, so we'll need to reload it here
